커뮤니티 댓글 달기 기능 구현
커뮤니티 화면에서 댓글 달기 기능 구현

+++ app/src/main/java/com/android/finder/dataobj/CommunityDetailDto.kt
@@ -4,7 +4,7 @@ data class CommunityDetailDto(
     val answerCount: Int,
     val answerHistDtos: List<CommentData>,
     val communityContent: String,
-    val communityId: Int,
+    val communityId: Long,
     val communityImgDtos: List<CommunityImgDto>,
     val communityMBTI: String,
     val communityTitle: String,
